refactor: Split block hooks
