+++ src/components/BlockProgress.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Progress, ProgressProps } from '@pancakeswap/uikit'
-import { useBlock } from 'state/hooks'
+import { useBlock } from 'state/block/hooks'
 
 interface BlockProgressProps extends ProgressProps {
   startBlock: number
