+++ src/state/block/hooks.ts
@@ -1,0 +1,27 @@
+import { useEffect } from 'react'
+import { useSelector } from 'react-redux'
+import { useAppDispatch } from 'state'
+import { simpleRpcProvider } from 'utils/providers'
+import { setBlock } from '.'
+import { State } from '../types'
+
+export const usePollBlockNumber = () => {
+  const dispatch = useAppDispatch()
+
+  useEffect(() => {
+    const interval = setInterval(async () => {
+      const blockNumber = await simpleRpcProvider.getBlockNumber()
+      dispatch(setBlock(blockNumber))
+    }, 6000)
+
+    return () => clearInterval(interval)
+  }, [dispatch])
+}
+
+export const useBlock = () => {
+  return useSelector((state: State) => state.block)
+}
+
+export const useInitialBlock = () => {
+  return useSelector((state: State) => state.block.initialBlock)
+}
