+++ src/views/Ifos/hooks/v1/useGetPublicIfoData.ts
@@ -2,7 +2,7 @@ import { useEffect, useState, useCallback } from 'react'
 import BigNumber from 'bignumber.js'
 import { BSC_BLOCK_TIME } from 'config'
 import { Ifo, IfoStatus, PoolIds } from 'config/constants/types'
-import { useBlock } from 'state/hooks'
+import { useBlock } from 'state/block/hooks'
 import { useLpTokenPrice } from 'state/farms/hooks'
 import { BIG_ZERO } from 'utils/bigNumber'
 import { multicallv2 } from 'utils/multicall'
