+++ src/views/Pools/components/PoolsTable/Cells/EndsInCell.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Flex, Link, Skeleton, Text, TimerIcon } from '@pancakeswap/uikit'
 import { getBscScanLink } from 'utils'
 import { Pool } from 'state/types'
-import { useBlock } from 'state/hooks'
+import { useBlock } from 'state/block/hooks'
 import Balance from 'components/Balance'
 import { useTranslation } from 'contexts/Localization'
 import { getPoolBlockInfo } from 'views/Pools/helpers'
