+++ src/views/Predictions/components/RoundCard/ExpiredRoundCard.tsx
@@ -4,7 +4,8 @@ import { useWeb3React } from '@web3-react/core'
 import { Box, BlockIcon, CardBody } from '@pancakeswap/uikit'
 import { useTranslation } from 'contexts/Localization'
 import { NodeRound, BetPosition, NodeLedger } from 'state/types'
-import { useBlock, useGetBetByEpoch } from 'state/hooks'
+import { useGetBetByEpoch } from 'state/hooks'
+import { useBlock } from 'state/block/hooks'
 import { formatBigNumberToFixed } from 'utils/formatBalance'
 import { getHasRoundFailed, getNetPayoutv2 } from '../../helpers'
 import { RoundResult } from '../RoundResult'
