+++ src/views/Predictions/components/RoundCard/LiveRoundCard.tsx
@@ -6,7 +6,8 @@ import { useTranslation } from 'contexts/Localization'
 import { NodeRound, NodeLedger, BetPosition } from 'state/types'
 import { BLOCK_PADDING } from 'state/predictions'
 import { formatBigNumberToFixed } from 'utils/formatBalance'
-import { useBlock, useGetLastOraclePrice } from 'state/hooks'
+import { useGetLastOraclePrice } from 'state/hooks'
+import { useBlock } from 'state/block/hooks'
 import BlockProgress from 'components/BlockProgress'
 import { formatUsdv2, getPriceDifference } from '../../helpers'
 import PositionTag from '../PositionTag'
