+++ src/views/Predictions/components/RoundCard/OpenRoundCard.tsx
@@ -4,7 +4,8 @@ import { CardBody, PlayCircleOutlineIcon, Button, useTooltip, ArrowUpIcon, Arrow
 import { useTranslation } from 'contexts/Localization'
 import { useAppDispatch } from 'state'
 import { BetPosition, NodeLedger, NodeRound } from 'state/types'
-import { useBlock, useGetIntervalBlocks } from 'state/hooks'
+import { useGetIntervalBlocks } from 'state/hooks'
+import { useBlock } from 'state/block/hooks'
 import { BLOCK_PADDING, fetchLedgerData } from 'state/predictions'
 import useToast from 'hooks/useToast'
 import CardFlip from '../CardFlip'
