+++ src/views/Predictions/index.tsx
@@ -3,7 +3,8 @@ import { useWeb3React } from '@web3-react/core'
 import { Helmet } from 'react-helmet-async'
 import { useMatchBreakpoints, useModal } from '@pancakeswap/uikit'
 import { useAppDispatch } from 'state'
-import { useGetPredictionsStatus, useInitialBlock, useIsChartPaneOpen } from 'state/hooks'
+import { useGetPredictionsStatus, useIsChartPaneOpen } from 'state/hooks'
+import { useInitialBlock } from 'state/block/hooks'
 import { initializePredictions } from 'state/predictions'
 import { PredictionStatus } from 'state/types'
 import usePersistState from 'hooks/usePersistState'
